requirements.txt
asgiref==3.7.2
crispy-bootstrap5==2023.10
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.2
django-bootstrap5==23.3
django-crispy-forms==2.1
django-environ==0.10.0
django-import-export==3.3.3
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
Pillow==10.0.0
python-decouple==3.8
PyYAML==6.0.1
reportlab==4.0.7
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.7.1
tzdata==2023.3
whitenoise==6.6.0
xlrd==2.0.1
xlwt==1.3.0
